fix(global): 세션 쿠키 도메인 적용을 위한 CookieSerializer 추가
